+++ .github/workflows/test.yml
@@ -22,7 +22,7 @@ jobs:
           python-version: ${{ matrix.py }}
       - uses: actions/checkout@v2
       # TODO: #48 Only install test requirements instead of all pinned in CI
-      - run: pip install -e . -r requirements.txt
+      - run: pip install -e . -r requirements.test.txt
       - run: pytest -v code_data
   docs:
     name: docs build
@@ -33,7 +33,7 @@ jobs:
         with:
           python-version: "3.10"
       - uses: actions/checkout@v2
-      - run: pip install -e . -r requirements.txt
+      - run: pip install -e . -r requirements.docs.txt
       - run: jupyter-book build docs --warningiserror --nitpick
       - name: Cat build logs
         run: cat docs/_build/html/reports/usage.log
