+++ .readthedocs.yaml
@@ -22,4 +22,4 @@ python:
    install:
       - method: pip
         path: .
-      - requirements: docs/requirements.txt
+      - requirements: requirements-docs.txt
