+++ code_data/test.py
@@ -2,12 +2,10 @@ from __future__ import annotations
 
 import dis
 import pathlib
-import pkgutil
 import sys
 import warnings
 from datetime import timedelta
 from dis import _get_instructions_bytes  # type: ignore
-from importlib.abc import Loader
 from types import CodeType
 from typing import Any, Iterable, Optional, cast
 
@@ -16,7 +14,8 @@ import pytest
 import rich.progress
 from hypothesis import HealthCheck, given, settings
 
-from code_data.line_mapping import (
+from . import from_code_data, to_code_data
+from .line_mapping import (
     USE_LINETABLE,
     LineMapping,
     bytes_to_items,
@@ -26,8 +25,7 @@ from code_data.line_mapping import (
     items_to_mapping,
     mapping_to_items,
 )
-
-from . import from_code_data, to_code_data
+from .module_codes import module_codes
 
 NEWLINE = "\n"
 
@@ -164,28 +162,6 @@ def test_generated(source_code):
         warnings.simplefilter("ignore")
         code = compile(source_code, "<string>", "exec")
     verify_code(code)
-
-
-def module_codes() -> Iterable[tuple[str, str, CodeType]]:
-    # In order to test the code_data, we try to get a sample of bytecode,
-    # by walking all our packages and trying to load every module.
-    # Note that although this doesn't require the code to be executable,
-    # `walk_packages` does require it, so this will ignore any modules
-    # which raise errors on import.
-
-    with warnings.catch_warnings():
-        # Ignore warning on find_module which will be deprecated in Python 3.12
-        # Worry about it later!
-        warnings.simplefilter("ignore")
-        for mi in pkgutil.walk_packages(onerror=lambda _name: None):
-            loader: Loader = mi.module_finder.find_module(mi.name)  # type: ignore
-            try:
-                code = loader.get_code(mi.name)  # type: ignore
-            except SyntaxError:
-                continue
-            if code:
-                source = loader.get_source(mi.name)  # type: ignore
-                yield mi.name, source, code
 
 
 def verify_code(code: CodeType) -> None:
