+++ requirements.test.txt
@@ -1,0 +1,27 @@
+#
+# This file is autogenerated by pip-compile with python 3.10
+# To update, run:
+#
+#    pip-compile --allow-unsafe --annotation-style=line --extra=test --output-file=requirements.test.txt pyproject.toml
+#
+attrs==21.4.0             # via hypothesis, pytest
+commonmark==0.9.1         # via rich
+exceptiongroup==1.0.0rc8  # via hypothesis
+hypothesis==6.48.2        # via hypothesmith
+hypothesmith==0.2.0       # via code_data (pyproject.toml)
+iniconfig==1.1.1          # via pytest
+lark-parser==0.12.0       # via hypothesmith
+libcst==0.4.5             # via hypothesmith
+mypy-extensions==0.4.3    # via typing-inspect
+packaging==21.3           # via pytest
+pluggy==1.0.0             # via pytest
+py==1.11.0                # via pytest
+pygments==2.12.0          # via rich
+pyparsing==3.0.9          # via packaging
+pytest==7.1.2             # via code_data (pyproject.toml)
+pyyaml==6.0               # via libcst
+rich==12.4.4              # via code_data (pyproject.toml)
+sortedcontainers==2.4.0   # via hypothesis
+tomli==2.0.1              # via pytest
+typing-extensions==4.3.0  # via libcst, typing-inspect
+typing-inspect==0.7.1     # via libcst
